+++ module/sheets/character-sheet.mjs
@@ -13,7 +13,7 @@ export class VermineCharacterSheet extends VermineActorSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
       classes: ["vermine2047", "sheet", "actor"],
-      template: "systems/vermine2047/templates/actor/actor-sheet.html",
+      template: "systems/vermine2047/templates/actor/actor-sheet.hbs",
       width: 600,
       height: 600,
       tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "features" }]
@@ -22,7 +22,7 @@ export class VermineCharacterSheet extends VermineActorSheet {
 
   /** @override */
   get template() {
-    return `systems/vermine2047/templates/actor/actor-character-sheet.html`;
+    return `systems/vermine2047/templates/actor/actor-character-sheet.hbs`;
   }
 
   /* -------------------------------------------- */