+++ module/sheets/npc-group.mjs
@@ -12,7 +12,7 @@ export class VermineGroupSheet extends VermineActorSheet {
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
       classes: ["vermine2047", "sheet", "actor"],
-      template: "systems/vermine2047/templates/actor/actor-sheet.html",
+      template: "systems/vermine2047/templates/actor/actor-sheet.hbs",
       width: 600,
       height: 600,
       tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "features" }]
@@ -21,7 +21,7 @@ export class VermineGroupSheet extends VermineActorSheet {
 
   /** @override */
   get template() {
-    return `systems/vermine2047/templates/actor/actor-${this.actor.type}-sheet.html`;
+    return `systems/vermine2047/templates/actor/actor-${this.actor.type}-sheet.hbs`;
   }
 
   /* -------------------------------------------- */